v0.69 PR1 (partial Phase 2) — split sound-effect, music-status, video-cancel
Plan G Phase 2 — generate.ts per-subcommand split. This PR establishes
the pattern with 3 small subcommands; remaining 10 (image, video,
speech, music, storyboard, thumbnail, background, video-status,
video-extend) come in follow-up PRs in subsequent sessions.

Why partial: full split is ~2,500 lines moved across 13 files —
mechanical but voluminous, doesn't fit a single session. Pattern
validation matters more than completeness for the first ship.

## Pattern (per file)

`packages/cli/src/commands/generate/<subcommand>.ts` shape:
1. Imports (relative paths bumped one level: `../utils/...`)
2. Library export: `executeXxx` + `ExecuteXxxOptions` / `ExecuteXxxResult`
   (when the subcommand has one — used by pipeline executor / manifest)
3. CLI export: `registerXxxCommand(parent: Command): void`
   wraps the original `parent.command("xxx")...action(...)` chain

`generate.ts` becomes a barrel: imports each `register*Command` and
calls `register*Command(generateCommand)`. Re-exports the library
functions for backward compat (so `from "./generate.js"` paths still
work in pipeline/executor.ts and other consumers).

## Splits in this PR

- `commands/generate/sound-effect.ts` (151 L) — ElevenLabs SFX
- `commands/generate/music-status.ts` (124 L) — Replicate status check
- `commands/generate/video-cancel.ts` (82 L) — Grok/Runway cancellation

`generate.ts`: 2,542 → 2,324 lines (−218 net via splits + thin wiring).
Each split file = original action body + executeXxx (when present),
imports adjusted for new depth. Behavior preserved (typecheck +
lint + 652 cli tests + 46 mcp tests + sync-counts all green).

## Out of scope (this PR)

- 10 remaining subcommand splits — future PRs
- Final generate.ts barrel collapse (~80 lines) — happens after all 13
- Manifest path updates (still imports from generate.ts re-export shim)
- v0.69.0 version bump (waits until Phase 2 fully done)
